pyproject.toml
[project]
name = "ai-content-generator"
version = "0.1.0"
description = "AI Content Generation Service for Marketplaces - Diploma Project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiogram>=3.2.0",
    "boto3>=1.34.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "diffusers>=0.24.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "pillow>=10.1.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "clip-interrogator>=0.6.0",
    "sentence-transformers>=2.2.0",
    "openai>=1.3.0",
    "aiofiles>=23.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "peft>=0.7.0",
    "optimum>=1.14.0",
    "tqdm>=4.64.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]